.boundary.method {phyclust}	R Documentation

Boundary Methods for Population Proportions

Description

Methods used in EM Algorithms to deal with boundary problems of population proportions, eta_k. The first element is the default value. This is a read-only object and the elemental order is followed in C.

Usage

.boundary.method

Format

A character vector contains implemented boundary methods in C.

Details

The boundary value 0 of the population proportions makes the log likelihood as -Inf. Since degeneracy of subpopulations can affect the maximizing processes in EM steps. This problem is usually caused by bad initializations, and may suggest that number of cluster K may be too large.

Two methods have been implemented when any eta_k less than the lower bound (1/N or 1e-16). The ADJUST (default) will adjust the eta_k to the lower bound, and the IGNORE will stop the iterations and return errors.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

.show.option, .init.procedure, .init.method, .EMControl, phyclust.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.boundary.method

## End(Not run)

[Package phyclust version 0.1-18 Index]