.edist.model {phyclust}	R Documentation

Evolution Distance Model

Description

Evolution distances based on certain evolution models as in ape. The implemented model is used in phyclust.edist and initializations of EM algorithms. The first element is the default value. This is a read-only object and the elemental order is followed in C.

Usage

.edist.model

Format

A character vector contains implemented evolution distances in C.

Details

This vector stores possible evolution distances implemented in phyclust. The default value is D_JC69 computed form the probability of JC69 model.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://thirteen-01.stat.iastate.edu/snoweye/phyclust/

See Also

.show.option, phyclust.edist.

Examples

## Not run: 
library(phyclust, quiet = TRUE)

.edist.model

## End(Not run)

[Package phyclust version 0.1-16 Index]