.em.method {phyclust}	R Documentation

EM Methods and Algorithms

Description

The varied EM algorithms are implemented in C. The first element is the default value. This is a read-only object and the elemental order is followed in C.

Usage

.em.method

Format

A character vector contains implemented EM algorithms in C.

Details

EM (default) stands for the standard EM algorithm, ECM stands for Expectation/Conditional Maximization algorithm, and AECM stands for Alternating ECM algorithm. The performance is roughly about AECM > EM ~ ECM which are dependent on the separations of data set.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

Dempster, A. and Laird, N. and Rubin, D. (1977) “Maximum Likelihood Estimation from Incomplete Data via the EM Algorithm”, Journal of the Royal Statistical Society Series B, 39:3, 1-38.

Meng, X.-L. and Rubin, D. (1993) “Maximum likelihood estimation via the ECM algorithm: A general framework”, Biometrika, 80:2, 511-567.

Meng, X.-L. and van Dyk, D. (1997) “The EM Algorithm — an Old Folk-song Sung to a Fast New Tune (with discussion)”, Journal of the Royal Statistical Society Series B, 59, 511-567.

See Also

.show.option, .EMC, .EMControl, phyclust.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.em.method

## End(Not run)

[Package phyclust version 0.1-18 Index]