.identifier {phyclust}	R Documentation

Identifiers for Evolution Models

Description

Identifiers for evolution models identify the Q_k matrix and evolution time t_k for subpopulations. The first element is the default value. This is a read-only object and the elemental order is followed in C.

Usage

.identifier

Format

A character vector contains implemented identifiers in C.

Details

Four major identifiers are implemented in C, EE, EV, VE, and VV. The first letter indicates the structure for Q_k matrix, and the second letter indicates the evolution time t_k for subpopulations. E and V indicate equal and varied for all subpopulations.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://thirteen-01.stat.iastate.edu/snoweye/phyclust/

See Also

.show.option, .EMC, .EMControl, phyclust.

Examples

## Not run: 
library(phyclust, quiet = TRUE)

.identifier

## End(Not run)

[Package phyclust version 0.1-16 Index]