.init.procedure {phyclust}	R Documentation

Initialization Procedures for EM Algorithms

Description

The varied initialization procedures are implemented in C. The first element is the default value. This is a read-only object and the elemental order is followed in C.

Usage

.init.procedure

Format

A character vector contains implemented initialization procedures in C.

Details

exhaustEM, emEM, RndEM, and RndpEM are implemented. Based on initialization states given by a initialization method, see .init.method for more information. These procedures will search a better starting states for final EM steps.

'exhaustEM'runs each initialization with EM steps until convergent, and pick the best one of the convergence as the return result.

'emEM'uses few short EM steps to improve initialization, then pick the best of initialization state for long EM steps, and returns the final result.

'RandEM'bases on initialization methods to generate initialization states, the number is equal to short EM steps, then pick the best of initialization state for long EM steps, and returns the final result.

'RandEM'bases on initialization methods to generate initialization states and run a fixed number of EM steps, until total steps exhaust short EM steps, then pick the best of initialization state for long EM steps, and returns the final result.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

Biernacki, C. and Celeux, G. and Govaert, G. (2003) “Choosing starting values for the EM algorithm for getting the highest likelihood in multivariate Gaussian mixture models”, Computational Statistics and Data Analysis, 41:3, 561-575.

Maitra, R. (2009) “Initializing partition-optimization algorithms”, IEEE/ACM Transactions on Computational Biology and Bioinformatics, 6:1, 114-157.

See Also

.show.option, .init.method, .EMControl, phyclust.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.init.procedure

## End(Not run)

[Package phyclust version 0.1-18 Index]