.label.method {phyclust}	R Documentation

Label Method

Description

An object stores label method for un-, semi-, and general semi- supervised clustering.. This is a read-only object and the elemental order is followed in C.

Usage

.label.method

Format

A character vector contains implemented evolution distances in C.

Details

This vector stores possible label methods implemented in phyclust. The default value is NONE for unsupervised clustering. SEMI is for semi-supervised clustering, and GENERAL is for general semi-supervised clustering. Only un- and semi-supervised clustering are implemented.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://thirteen-01.stat.iastate.edu/snoweye/phyclust/

See Also

phyclust.

Examples

## Not run: 
library(phyclust, quiet = TRUE)

.label.method

## End(Not run)

[Package phyclust version 0.1-16 Index]