.show.option {phyclust}	R Documentation

Show Available Options

Description

This function show available options for functions in phyclust.

Usage


  .show.option()

Details

This function show some available options for functions in phyclust. They are used in .EMControl, phyclust, ... etc, and options are stored in several objects separately. They will be passed into C, so the elemental order are important. Basically, they are all read-only objects.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

.boundary.method, .code.type, .edist.model, .em.method, .EMC, .EMControl, .identifier, .init.method, .init.procedure, .nucleotide, .snp, .substitution.model,

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.show.option()

## End(Not run)

[Package phyclust version 0.1-18 Index]