standard.code {phyclust}	R Documentation

Standard Codes and ids for Nucleotides, SNPs, Codon, Amino Acid and Genetic Code

Description

Standard codes and ids for nucleotides, SNPs, codon, amino acid and genetic code. All objects are used to transfer data. These are read-only objects and the elemental order is followed in C.

Usage


.nucleotide
.snp
.codon
.amino.acid
.genetic.code

Format

All objects are data frames containing ids and codes.

Details

Note: All ids are coding started from 0.

Nucleotides, A, G, C, T, and - have codes 0, 1, 2, 3, and 4 where "-" is for gaps. SNPs, 1, 2, and - have code codes 0, 1, and 2.

These are objects are in data frames unlike other internal objects due to heavily used in processing data. The original data should be transfered to numerical codes in order to be passed to C codes. In C codes, we use integers, 0, 1, 2, ..., for coding nucleotides or SNPs and so on.

Now, models and methods are implemented only for .nucleotide and .snp. Other objects are leaved for further extension.

Data frames use factor formats as the default, and as.character is the way to transfer to the characters.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

.show.option, code2nid, nid2code, snp2sid, sid2snp.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.nucleotide
.snp
.codon
.amino.acid
.genetic.code
.missing.code

## End(Not run)

[Package phyclust version 0.1-18 Index]