.substitution.model {phyclust}	R Documentation

Substitution Models for Mutation Processes

Description

An object stores substitution models for mutation processes for Continuous Time Markov Chain theory. This is a read-only object and the elemental order is followed in C.

Usage

.substitution.model

Format

A data frame contains two character vectors, mode and code.type.

Details

This data frame indicates substitution models implemented in C.

'model': names of substitution models for mutations.

'code.type': code types of substitution models, either NUCLEOTIDE or SNP.

The major models are:

Model 	 Author and Publication 	 Parameter

JC69 	 Jukes and Cantor 1969. 	 t

K80 	 Kimura 1980. 	 kappa, t

F81 	 Felsenstein 1981. 	 pi, t

HKY85 	 Hasegawa, Kishino, and Yano 1985. 	 kappa, pi, t

Other models starting with E_ use empirical frequencies for equilibrium probabilities.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://thirteen-01.stat.iastate.edu/snoweye/phyclust/

Jukes, T. H. and Cantor, C. R. (1969) “Evolution of Protein Molecules”, Mammalian Protein Metabolism, 3, 21-132

Kimura, M. (1980) “A Simple Method for Estimating Evolutionary Rates of Base Substitutions through Comparative Studies of Nucleotide Sequences”, Journal of Molecular Evolution, 16, 111-120

Felsenstein, J. (1981) “Evolutionary Trees from DNA Sequences: A Maximum Likelihood Approach”, Journal of Molecular Evolution, 17, 368-376

Hasegawa, M. and Kishino, H. and Yano, T. (1985) “Dating of the Human-Ape Splitting by a Molecular Clock of Mitochondrial DNA”, Journal of Molecular Evolution, 22:2, 160-174

See Also

.show.option, .code.type, .identifier, .EMControl, phyclust.

Examples

## Not run: 
library(phyclust, quiet = TRUE)

.substitution.model

## End(Not run)

[Package phyclust version 0.1-16 Index]