bootstrap.seq.data {phyclust}	R Documentation

Bootstrap a seq.data from a Fitted Model.

Description

This function simplifies the bootstrap function bootstrap.star.trees.seq(), and only return a list object with class seq.data.

Usage


bootstrap.seq.data(ret.phyclust, min.n.class = 1)

Arguments

ret.phyclust
a phyclust object in Class phyclust.

min.n.class
minimum number of sequences for a cluster.

Details

ret.phyclust is a phyclust object in Class phyclust which is usually fitted by phyclust, or returned by phyclust.m.step.

min.n.class is a boundary condition to avoid degenerate clusters when some population proportions, eta_k, are small in the fitted model.

Value

Return an object in Class seq.data as the result from read.*().

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

phyclust, bootstrap.star.trees, Class seq.data.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

set.seed(1234)
EMC.1 <- .EMC
EMC.1$EM.iter <- 1
# the same as EMC.1 <- .EMControl(EM.iter = 1)

ret.1 <- phyclust(seq.data.toy$org, 2, EMC = EMC.1)
(ret.all <- bootstrap.seq.data(ret.1))

## End(Not run)

[Package phyclust version 0.1-18 Index]