bootstrap.star.trees {phyclust}	R Documentation

Bootstrap a Star Tree from a Fitted Model.

Description

This function bootstraps a star tree from a model fitted by phyclust. Each cluster corresponds to a star tree and a center sequence where sequences will evolve from. This function is called by bootstrap.star.trees.seq to generate sequences. The fitted model can be varied in .identifier.

Usage


bootstrap.star.trees(ret.phyclust, min.n.class = 1)

Arguments

ret.phyclust
a phyclust object in Class phyclust.

min.n.class
minimum number of sequences for a cluster.

Details

ret.phyclust is a phyclust object in Class phyclust which is usually fitted by phyclust, or returned by phyclust.m.step.

min.n.class is a boundary condition to avoid degenerate clusters when some population proportions, eta_k, are small in the fitted model.

Value

Return a list containing K star trees according to ret.phyclust.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

phyclust, bootstrap.seq, bootstrap.star.trees.seq.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

set.seed(1234)
EMC.1 <- .EMC
EMC.1$EM.iter <- 1
# the same as EMC.1 <- .EMControl(EM.iter = 1)

ret.1 <- phyclust(seq.data.toy$org, 2, EMC = EMC.1)
ret.trees <- bootstrap.star.trees(ret.1)

## End(Not run)

[Package phyclust version 0.1-18 Index]