gen.star.tree {phyclust}	R Documentation

Generate a Rooted Tree with a Star Shape

Description

Generate a rooted tree with a star shape based on a sequence calls of several functions.

Usage


gen.star.tree(N, total.height = 1)

Arguments

N
number of leaves.

total.height
total tree height.

Details

A tree with a star shape means that all internal branches are 0 and all leaf branches are equal.

This function combining with gen.seq.HKY or gen.seq.SNP is used in simulation studies and bootstrap tree samples

Value

Return a rooted tree in Class phylo with a star shape.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

ms. as.star.tree, get.rooted.tree.height. rescale.rooted.tree. as.phylo, plot.phylo.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

ret.star <- gen.star.tree(5)
plot(ret.star, type = "u")

## End(Not run)

[Package phyclust version 0.1-18 Index]