phyclust.Pt {phyclust}	R Documentation

Transition Probabilities of phyclust Given Time

Description

This computes transition probabilities of phyclust given time.

Usage


phyclust.Pt(Q, Tt, substitution.model = .substitution.model$model[1])

Arguments

Q
a list according to the substitution model.

Tt
total evolution time, t.

substitution.model
substitution model.

Details

The major models for Q are:

Model 	 Author and Publication 	 Parameter

JC69 	 Jukes and Cantor 1969. 	 t

K80 	 Kimura 1980. 	 kappa, t

F81 	 Felsenstein 1981. 	 pi, t

HKY85 	 Hasegawa, Kishino, and Yano 1985. 	 kappa, pi, t

A list of Q should contains pi, kappa based on substitution models and code types. Tt may be separately stored. Depending on identifiers, Qs can be composite to a QA, Q matrix array.

Value

A list with class Pt will be returned containing several elements as the following:

'Pt'
a transition probability matrix.

'log.Pt'
a log transition probability matrix.

'H'
a negative entropy, diag(Pt %*% t(log.Pt)).

ToDo(s)

vectorize Tt for repeated computation in C.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

.substitution.model, phyclust, phyclust.em.step.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

Tt <- 0.5

Q <- list(pi = c(0.25, 0.25, 0.25, 0.25), kappa = 0.5)
phyclust.Pt(Q, Tt, "HKY85")

Q <- list(pi = c(0.5, 0.5), kappa = 0.5)
phyclust.Pt(Q, Tt, "SNP_JC69")

## End(Not run)

[Package phyclust version 0.1-18 Index]