phyclust.se {phyclust}	R Documentation

The Main Function of phyclust for Sequencing Error Models

Description

The phyclust.se is an advanced function of phyclust. The phyclust.se implements finite mixture models for sequence data with sequencing errors. The same as phyclust, the mutation processes are modeled by evolution processes based on Continuous Time Markov Chain theory.

Usage


phyclust.se(X, K, EMC = .EMC, manual.id = NULL, byrow = TRUE)

Arguments

X
nid/sid matrix with N rows/sequences and L columns/sites.

K
number of clusters.

EMC
EM control.

manual.id
manually input class ids.

byrow
advanced option for X, default = TRUE.

Details

phyclust.se considers mutations with sequencing error, but only for NUCLEOTIDE data and only the EM algorithm is implemented. Currently, phyclust.se implements the following steps:

1 	 assume non-sequencing errors as the phyclust does.

2 	 use the initialization as the phyclust does.

3 	 run the phyclust to find the non-sequencing error MLE's.

4 	 initial by the results of phyclust.

5 	 update all parameters including probabilities of sequencing errors.

See the help page of phyclust for the explanations of input arguments since both functions share the same arguments. Note that the underling model assumptions are very different of both functions.

Value

A list with class phyclust will be returned containing several elements as described in phyclust. But, the phyclust.se returns extra parameters for sequencing errors, and they are shown in the following:

'SE'
a list returning parameters of sequencing error models, including:

'type' 	 'TRUE' for modeling sequencing errors.

'model' 	 'CONVOLUTION', the only model implemented.

'constant' 	 the constrained constant for sequencing errors.

'f.err' 	 probability matrix, each row sums to 1.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

.EMC, .EMControl, phyclust.se, phyclust.se.update.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

X <- seq.data.toy$org

set.seed(1234)
(ret.1 <- phyclust.se(X, 3))

## End(Not run)

[Package phyclust version 0.1-18 Index]