phyclust.update {phyclust}	R Documentation

Update phyclust Results

Description

This function will run the EM algorithm on initial parameters specified by users or from other initial procedures. All parameters (Eta, Mu, Q, ...) in this function will be updated.

Usage


phyclust.update(X, EMC = .EMC, ret.phyclust = NULL, K = NULL,
    Eta = NULL, Mu = NULL, pi = NULL, kappa = NULL, Tt = NULL,
    label = NULL, byrow = TRUE)

Arguments

X
nid/sid matrix with N rows/sequences and L columns/sites.

EMC
EM control.

ret.phyclust
an object with the class phyclust.

K
number of clusters.

Eta
proportion of subpopulations, eta_k, length = K, sum to 1.

Mu
centers of subpopulations, dim = K*L, each row is a center.

pi
equilibrium probabilities, each row sums to 1.

kappa
transition and transversion bias.

Tt
total evolution time, t.

label
label of sequences for semi-supervised clustering.

byrow
advanced option for X, default = TRUE.

Details

This function is equivalent to run exhaustEM on one specified initial parameters, and no initial procedure is involved. While this function is a little bit different to run phyclust with manual.id where Mu will be reestimated as the new initials. Simply speaking, this function only runs the EM algorithm given the initial parameters.

All the input arguments are the same as the inputs of the functions phyclust and phyclust.em.step.

Value

This function returns an object with class phyclust.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

phyclust, find.best, phyclust.se, phyclust.se.update.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

set.seed(1234)
EMC.1 <- .EMC
EMC.1$EM.iter <- 1
# the same as EMC.1 <- .EMControl(EM.iter = 1)
X <- seq.data.toy$org

(ret.1 <- phyclust(X, 2, EMC = EMC.1))
(ret.2 <- phyclust.update(X, ret.phyclust = ret.1))

## End(Not run)

[Package phyclust version 0.1-18 Index]