plotgaps {phyclust}	R Documentation

Gaps Plots of Sequences for Visual Comparisons

Description

This function provides gaps plots of data set to identify regions where gaps enriched. The plot show the proportions of context by sites and the diverse may be caused by mutations, sequencing errors, or alignment errors.

Usage


plotgaps(X, code.type = .code.type[1], main = "Gaps Plot",
    xlab = "Sites", ylab = "Proportion", ...)

Arguments

X
numerical data matrix with N rows/sequences and L columns/sites.

code.type
either "NUCLEOTIDE" (default) or "SNP".

main
main label, default = "Gaps Plot".

xlab
x-axis label, default = "Sites".

ylab
y-axis label, default = "Proportion".

...
other options passed to plot.

Details

Proportions of gaps will be drawn.

Value

A gaps plot will be drawn.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

plotdots.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

# For nucleotide
set.seed(1234)
X <- seq.data.toy$org
X[sample(c(T, F), length(X), replace = TRUE, prob = c(0.05, 0.95))] <-
  .nucleotide$nid[.nucleotide$code == "-"]
plotgaps(X)

## End(Not run)

[Package phyclust version 0.1-18 Index]