print.object {phyclust}	R Documentation

Functions for Printing or Summarizing Objects According to Classes

Description

Several classes are declared in phyclust, and these are functions to print and summary objects.

Usage


## S3 method for class 'baseml'
print(x, ...)
## S3 method for class 'ms'
print(x, ...)
## S3 method for class 'phyclust'
print(x, digits = max(4, getOption("digits") - 3), ...)
## S3 method for class 'Pt'
print(x, ...)
## S3 method for class 'RRand'
print(x, digits = max(4, getOption("digits") - 3), ...)
## S3 method for class 'seq.data'
print(x, ...)
## S3 method for class 'seqgen'
print(x, ...)
## S3 method for class 'phyclust'
summary(object, ...)

Arguments

x
an object with the class attributes.

digits
for printing out numbers.

object
an object with the class attributes.

...
other possible options.

Details

These are useful functions for summarizing and debugging.

For ms, seqgen, and paml.baseml, it will show the result as standalone versions on the STDOUT out with line by line.

For other functions, they only show summaries of objects. Use names or str to explore the details.

Value

The results will cat or print on the STDOUT by default.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

ms, paml.baseml, phyclust, phyclust.Pt, RRand, seqgen.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

# Functions applied by directly type the names of objects.

## End(Not run)

[Package phyclust version 0.1-18 Index]