.EMC {phyclust}	R Documentation

EM Control

Description

An default template object stores controlling options for phyclust to perform EM algorithms. This object combines all other read-only objects and more required options for EM algorithms. This is essential for phyclust and other related functions.

Usage

.EMC

Format

A list contains all controlling options

Details

A list created by .EMControl contains all controlling options for EM algorithms. This list will be directly passed to C codes and control the every things of EM algorithms.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

.show.option, .EMControl, phyclust.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.EMC

## End(Not run)

[Package phyclust version 0.1-18 Index]