.code.type {phyclust}	R Documentation

Code Types of Dataset and Substitution Models

Description

Indicate the types of codes for datasets and functions. The first element is the default value. This is a read-only object and the elemental order is followed in C.

Usage

.code.type

Format

A character vector contains implemented code types in C.

Details

Two possible types are implemented, "NUCLEOTIDE" (default) and "SNP", used in data transfers and indicating substitution models.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

See Also

.show.option, .substitution.model, .EMControl, phyclust.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.code.type

## End(Not run)

[Package phyclust version 0.1-18 Index]