.init.method {phyclust}	R Documentation

Initialization Methods for EM Algorithms

Description

The varied initialization methods are implemented in C. The first element is the default value. This is a read-only object and the elemental order is followed in C.

Usage

.init.method

Format

A character vector contains implemented initialization methods in C.

Details

randomMu, NJ, randomNJ, PAM, K-Medoids and manualMu are implemented where the codes for the NJ are modified from ape, and the codes for the PAM method are modified from cluster. These methods are only provide initializations for EM algorithms.

'randomMu'randomly picks centers and assigns all sequences near by the center according an evolution distance.

'NJ'bases on a neighbor-joining tree and partitions the tree by the long branches into subtrees to form clusters.

'randomNJ'randomly partition a neighbor-joining tree into subtrees to form clusters.

'PAM'uses the partition around medoids algorithm to locate the centers of dataset.

'K-Medoids'performs K-Means types algorithms to randomly and roughly locate centers and form clusters.

'manualMu'requires a vector containing class ids for all sequences.

Author(s)

Wei-Chen Chen wccsnow@gmail.com

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

Saitou, N. and Nei, M. (1987), “The Neighbor-Joining Method: A New Method for Reconstructing Phylogenetic Trees”, Molecular Biology and Evolution, 4:4, 406-425.

Kaufman, L. and Rousseeuw, P.J. (1990), Finding Groups in Data: An Introduction to Cluster Analysis, Wiley.

Theodoridis, S. and Koutroumbas, K. (2006), Pattern Recognition, 3rd ed., pp. 635.

See Also

.show.option, .init.procedure, .EMControl, phyclust.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

.init.method

## End(Not run)

[Package phyclust version 0.1-18 Index]