data.phylip.crohn {phyclust}	R Documentation

Crohn's Disease SNP Dataset in the phylip Format

Description

Crohn's disease dataset is published by Hugot, et al. (2001).

Format

A text file in phylip format is stored in the data subdirectory.

Details

Crohn's disease dataset is used to perform haplotype grouping used in Tzeng's paper (2005).

Totally, 1102 haplotypes/SNP sequences and 8 sites.

Author(s)

Hugot, J.P., et al. (2001).

References

Phylogenetic Clustering Website: http://snoweye.github.io/phyclust/

Hugot, J.P., et al. (2001) “Association of NOD2 Leucine-Rich Repeat Variants with Susceptibility to Crohn's Disease”, Nature, 411, 599-603.

Tzeng, J.Y. (2005) “Evolutionary-Based Grouping of Haplotypes in Association Analysis”, Genetics Epidemiology, 28, 220-231. http://www4.stat.ncsu.edu/~jytzeng/software.php

See Also

read.phylip.

Examples


## Not run: 
library(phyclust, quiet = TRUE)

data.path <- paste(.libPaths()[1], "/phyclust/data/crohn.phy", sep = "")
# edit(file = data.path)
my.snp <- read.phylip(data.path, code.type = "SNP")
str(my.snp)

## End(Not run)

[Package phyclust version 0.1-18 Index]